backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "email-validator>=2.3.0",
    "faker>=40.8.0",
    "fastapi>=0.135.1",
    "ollama>=0.6.1",
    "openai>=2.26.0",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic[email]>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
    "tabulate>=0.10.0",
    "uvicorn>=0.41.0",
    "xai-sdk>=1.8.1",
]
